Update tugas alpro 1 & 2

+++ praktikum_2/1_InOut/luasKelLayang.c
@@ -1,7 +1,7 @@
 /*Nama File 	: luasKelLayang.c*/
 /*Deskripsi 	: Menghitung keliling dan luas layang-layang*/
 /*Pembuat   	: 24060124130069-Muhammad Fikri*/
-/*Tgl Pembuatan	: 28 Februari 2025 07.50 */
+/*Tgl Pembuatan	: 1 Maret 2025 08.00 */
 
 #include <stdio.h> /*header file*/
 
@@ -9,14 +9,14 @@
 int main()
 {
     /*Kamus*/
-    int s1, s2, d1, d2, keliling, luas;
+    float s1, s2, d1, d2, keliling, luas;
 
     /*Algoritma*/
     scanf("%d %d %d %d", &s1, &s2, &d1, &d2);
     keliling = 2 * (s1 + s2);
     luas = (d1 * d2) / 2;
-    printf("Keliling : %d\n", keliling);
-    printf("Luas : %d\n", luas);
+    printf("Keliling : %f\n", keliling);
+    printf("Luas : %f\n", luas);
 
     return 0;
 }
